setup/requirements.txt
neo4j==5.12.0
numpy==1.25.2
pandas==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
